# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cue API
  version: 0.1.0
  description: |
    HTTP contract between cue-api and cue-ios. Source of truth for endpoints,
    request/response shapes, and status codes. See `docs/architecture.md` and
    `docs/specs/` for design rationale.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.cue.app
    description: Production (placeholder — not yet deployed)
...
paths:
...
  /assistant/link:
    get:
      tags: [assistant]
      summary: Current Telegram link status
      description: |
        Reports whether the signed-in user has a linked Telegram chat. Drives the
        Settings "Connect Telegram" row. See
        `docs/specs/telegram-ai-assistant.md`.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current link state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramLinkStatus'
        '401':
          description: Missing, invalid, or expired bearer token.
    post:
      tags: [assistant]
      summary: Redeem a linking code
      description: |
        App-initiated binding. Burns the single-use nonce the bot issued and
        upserts the `TelegramLink` binding this user to the Telegram chat, then
        returns the resulting link state (`linked: true`).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkTelegramRequest'
      responses:
        '200':
          description: Linking succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramLinkStatus'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '422':
          description: |
            The nonce is invalid, expired, or already burned (M3). Returns a typed
            `{ code: 'INVALID_LINK_CODE', message }` body so iOS can branch a
            persistent bad-code state vs a transient error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidLinkCodeBody'
    delete:
      tags: [assistant]
      summary: Revoke the Telegram link
      description: |
        Disconnects the signed-in user's Telegram chat. Idempotent — succeeds
        with `linked: false` whether or not a link was present.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Link revoked (or already absent).
          content:
            application/json:
              schema:
                type: object
                required: [linked]
                properties:
                  linked:
                    type: boolean
                    enum: [false]
        '401':
          description: Missing, invalid, or expired bearer token.
  /assistant/parse:
    post:
      tags: [assistant]
      summary: Parse natural language into a task draft
      description: |
        Parses one line of natural language into a structured task draft (D4) via
        the assistant AI connector WITHOUT creating a task. JWT-guarded. The iOS
        quick-create well pre-fills its fields from the returned `TaskDraftDTO`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseTaskRequest'
      responses:
        '200':
          description: The extracted task draft.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDraftDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
...
  /users/me/devices:
    post:
      tags: [users]
      summary: Register an APNs device token
      description: |
        Registers (upserts) an APNs device token for the authed user. Idempotent —
        re-posting a known token re-claims/refreshes the row rather than
        duplicating. Returns 201 with the `DeviceDTO`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device registered (or refreshed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
  /users/me/devices/{token}:
    delete:
      tags: [users]
      summary: Unregister an APNs device token
      description: |
        Unregisters an APNs token owned by the authed user (scoped to the caller).
        Responds 200 `{ ok: true }` (JSON body, not 204, matching the repo's delete
        convention); 404 when the token is not owned by the caller.
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: The APNs device token to unregister.
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '401':
          description: Missing, invalid, or expired bearer token.
        '404':
          description: No such token owned by the caller.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RecurrenceRuleInput:
      type: object
      description: |
        Recurrence config payload sent inside create/update task and task-group
        request bodies — the same inline shape as RecurrenceConfig (normalized and
        stored verbatim as JSONB). Cross-field: COUNT requires `count`; UNTIL_DATE
        requires `endDate`.
      required: [frequency]
      properties:
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        interval:
          type: integer
          minimum: 1
          default: 1
        byWeekday:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
        byMonthDay:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 31
        byMonth:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 12
        bySetPos:
          type: array
          items:
            type: integer
            enum: [1, 2, 3, 4, -1]
          description: |
            MONTHLY ordinal selectors paired with `byWeekday` ("first Monday",
            "last Friday"). Values 1..4 or -1.
        monthlyAnchor:
          type: string
          enum: [FIRST_WORKDAY, LAST_WORKDAY, DAY_BEFORE_LAST_WORKDAY]
          description: |
            MONTHLY working-day anchor (first / last / day-before-last workday).
            Mutually exclusive with the `bySetPos`/`byMonthDay` axes.
        endType:
          type: string
          enum: [NEVER, UNTIL_DATE, COUNT]
          default: NEVER
        endDate:
          type: string
          format: date
          description: Required when endType = UNTIL_DATE.
        count:
          type: integer
          minimum: 1
          description: Required when endType = COUNT.
...
    LinkTelegramRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          description: |
            Single-use linking nonce the bot issued in the "Connect Telegram"
            prompt. The iOS app receives it via the universal link (or manual
            paste) and redeems it over the authenticated session.
    TelegramLinkStatus:
      type: object
      required: [linked, telegramUsername, linkedAt]
      description: |
        Telegram link state returned by `GET`, `POST`, and `DELETE
        /assistant/link`. When not linked, `telegramUsername` and `linkedAt` are
        null.
      properties:
        linked:
          type: boolean
        telegramUsername:
          type: [string, 'null']
          description: Telegram @handle, when known (may be null even when linked).
        linkedAt:
          type: [string, 'null']
          format: date-time
          description: ISO-8601 timestamp the link was established, or null when unlinked.
    OkResponse:
      type: object
      required: [ok]
      description: |
        Bare success acknowledgement `{ ok: true }` for endpoints that affect a
        resource without returning it (matches the repo's delete convention).
      properties:
        ok:
          type: boolean
          enum: [true]
...
    DevicePlatform:
      type: string
      enum: [ios, ipados, macos]
      description: The platform an APNs device token was issued on.
    RegisterDeviceRequest:
      type: object
      description: |
        Registers (upserts) an APNs device token for the current user. Idempotent —
        a re-posted known token re-claims/refreshes the row rather than duplicating.
      required: [token, platform]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 512
          description: The opaque APNs device token issued to the client.
        platform:
          $ref: '#/components/schemas/DevicePlatform'
    DeviceDTO:
      type: object
      description: A registered device. Returned by POST /users/me/devices. Omits the owning userId.
      required: [id, token, platform, lastSeenAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
          description: The opaque APNs device token.
        platform:
          $ref: '#/components/schemas/DevicePlatform'
        lastSeenAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ParseTaskRequest:
      type: object
      description: |
        One line of natural language to parse into a structured task draft (D4),
        WITHOUT creating a task. Bounded so a stray paste cannot drive an unbounded
        model call.
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
          example: 'gym every monday and wednesday at 7am'
    TaskDraftDTO:
      type: object
      description: |
        The structured task draft the assistant extracted from one line of natural
        language (D4). DRAFT only — no task is created; the iOS quick-create well
        pre-fills from this and the user confirms/edits before a real POST /tasks.
      required: [title]
      properties:
        title:
          type: string
          example: 'gym'
          description: Extracted task title, with scheduling words stripped.
        start:
          type: string
          format: date-time
          description: ISO-8601 start resolved in the user timezone; omitted for a timeless todo.
        durationMinutes:
          type: integer
          minimum: 1
          example: 60
          description: Present only when the text implied a length.
        recurrence:
          $ref: '#/components/schemas/RecurrenceRuleInput'
        groupId:
          type: string
          format: uuid
          description: Id of an EXISTING group the draft was matched to; omitted when none applied.
...
    InvalidLinkCodeBody:
      type: object
      description: |
        Typed 422 body returned by POST /assistant/link on an invalid, expired, or
        already-burned nonce (M3) so iOS branches a persistent bad-code state vs a
        transient error.
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_LINK_CODE]
        message:
          type: string
          example: 'This linking code is invalid, expired, or already used.'
        statusCode:
          type: integer
          example: 422
    ValidationErrorBody:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
          example: Bad Request
  responses:
    ValidationError:
      description: Request DTO failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
    OkResponse:
      description: 'Success acknowledgement `{ ok: true }`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'